refactor metadata
