+++ reporter/test/test_osm.py
@@ -9,10 +9,6 @@ from reporter.utilities import LOGGER
 from reporter.osm import (
     load_osm_document,
     extract_shapefile,
-    short_version,
-    latest_xml_metadata_file,
-    metadata_file,
-    metadata_files,
     check_string)
 from reporter.test.helpers import FIXTURE_PATH
 
@@ -57,49 +53,6 @@ class OsmTestCase(LoggedTestCase):
         file_time2 = os.path.getmtime(file_path)
         message = 'load_osm_document cache test failed.'
         self.assertEqual(file_time, file_time2, message)
-
-    def test_latest_xml_metadata_file(self):
-        """Test the maximum version available of an XML keyword file."""
-        self.assertTrue(latest_xml_metadata_file('building-points') == 3.2)
-        self.assertTrue(latest_xml_metadata_file('buildings') == 3.2)
-        self.assertTrue(latest_xml_metadata_file('roads') == 3.2)
-
-    def test_metadata_file(self):
-        """Test we get the good metadata file."""
-        file_suffix = metadata_file('keywords', '3.1', 'fake_lang', 'roads')
-        self.assertEqual('-en.keywords', file_suffix)
-
-        file_suffix = metadata_file('keywords', None, 'fr', 'roads')
-        self.assertEqual('-fr.keywords', file_suffix)
-
-        file_suffix = metadata_file('xml', '3.3', 'fake_lang', 'roads')
-        self.assertEqual('-3.2-en.xml', file_suffix)
-
-        file_suffix = metadata_file('xml', '3.2', 'en', 'roads')
-        self.assertEqual('-3.2-en.xml', file_suffix)
-
-    def test_metadata_files(self):
-        """Test we get all metadata files."""
-        metadata = metadata_files('3.2', 'en', 'roads', 'test')
-        expected_metadata = {'test.xml': '-3.2-en.xml'}
-        self.assertDictEqual(expected_metadata, metadata)
-
-        metadata = metadata_files(None, 'fr', 'roads', 'test')
-        expected_metadata = {
-            'test.xml': '-3.2-en.xml',
-            'test.keywords': '-fr.keywords'
-        }
-        self.assertDictEqual(expected_metadata, metadata)
-
-        metadata = metadata_files('3.1', 'en', 'roads', 'test')
-        expected_metadata = {'test.keywords': '-en.keywords'}
-        self.assertDictEqual(expected_metadata, metadata)
-
-    def test_short_version(self):
-        """Test the inasafe version."""
-        self.assertEquals(short_version('3.2.0.dev-dbb84de'), 3.2)
-        self.assertEquals(short_version('3.2.0'), 3.2)
-        self.assertEquals(short_version('3.2'), 3.2)
 
     def test_extract_shapefile(self):
         """Test the roads to shp converter."""
